{{cookiecutter.project_slug}}/backend/requirements/common.txt
beautifulsoup4
boto
celery
Django==1.11
django-cors-headers
django-debug-toolbar
django-environ
django-filter
django-hijack
django-htmlmin
django-parler
django-polymorphic
djangorestframework
django-storages
django-webpack-loader
django-widget-tweaks
drf-extensions
pillow
psycopg2
whitenoise
